fix ipad layout
